refactor: error handling in install script

+++ install.sh
@@ -1,15 +1,13 @@
 #!/bin/bash
 
-pre_install() {
+configure_dots() {
     pacman -S git --needed base-devel --noconfirm || exit
-}
 
-post_install() {
     git clone https://aur.archlinux.org/yay.git || exit
     cd yay || exit
     makepkg -si --noconfirm || exit
     cd .. && rm -rf yay || exit
-    yay -S neofetch skippy-xd cava ranger mpv rofi wezterm zsh feh xorg-xrandr neovim flameshot notify-osd discord chromium xclip pavucontrol ttf-roboto-mono-nerd thunar lxappearance fzf networkmanager exa bat github-cli zsh-autosuggestions zsh-syntax-highlighting meson ninja uthash libconfig nodejs npm python-pip libsixel imagemagick  --noconfirm || exit
+    yay -S neofetch skippy-xd cava ranger mpv rofi wezterm zsh feh xorg-xrandr neovim flameshot notify-osd discord chromium xclip pavucontrol ttf-roboto-mono-nerd thunar lxappearance fzf networkmanager exa bat github-cli zsh-autosuggestions zsh-syntax-highlighting meson ninja uthash libconfig nodejs npm python-pip libsixel imagemagick qt5-quickcontrols2-git qt5-graphicaleffects-git qt5-svg-git --noconfirm || exit
 
     if [ -e /etc/systemd/system/display-manager.service ]; then 
         default_target=$(basename $(readlink -f /etc/systemd/system/display-manager.service))
@@ -22,35 +20,46 @@ post_install() {
       sudo systemctl enable sddm.service || exit
     fi
 
-    yay -S qt5‑graphicaleffects qt5‑quickcontrols2 qt5‑svg || exit
-
     git clone https://github.com/fdev31/picom || exit
-    cd picom || exit
+    cd picom 
     meson setup --buildtype=release . build || exit
     ninja -C build || exit
     sudo cp build/src/picom /usr/bin/picom || exit
-    cd .. && rm -rf picom || exit
-}
+    cd .. 
+    rm -rf picom || exit
 
-configure_dots() {
-    git clone https://github.com/carrotshniper21/dotfiles || exit
-    cd dotfiles || exit
+    if [ -d ./dotfiles ]; then 
+      rm -rf ./dotfiles
+      git clone https://github.com/carrotshniper21/dotfiles || exit
+    else
+      git clone https://github.com/carrotshniper21/dotfiles || exit
+    fi
+
+    cd dotfiles
 
     chsh $USER -s /bin/zsh
-    rm $HOME/.bash*
+    
+    if [ -e $HOME/.bash* ]; then 
+      rm $HOME/.bash*
+    fi
 
     echo "(1/8) Configuring backgrounds..."
     cp -r backgrounds $HOME/.backgrounds
 
     echo "(2/8) Installing lunarvim..."
     bash <(curl -s https://raw.githubusercontent.com/LunarVim/LunarVim/release-1.3/neovim-0.9/utils/installer/install.sh)
-    sudo cp .local/bin/lvim /usr/bin/lvim || exit
+    sudo cp .local/bin/lvim /usr/bin || exit
 
     echo "(3/8) Copying configuration files..."
     cp -r config/* $HOME/.config || exit
 
     echo "(4/8) Installing GTK-3 theme..."
-    git clone https://github.com/jmattheis/gruvbox-dark-gtk ~/.themes/gruvbox-dark-gtk || exit
+    if [ -d $HOME/.themes ]; then 
+      rm -rf .themes
+      git clone https://github.com/jmattheis/gruvbox-dark-gtk $HOME/.themes/gruvbox-dark-gtk || exit
+    else 
+      git clone https://github.com/jmattheis/gruvbox-dark-gtk $HOME/.themes/gruvbox-dark-gtk || exit
+    fi
 
     echo "(5/8) Copying desktop configuration files..."
     cp desktop/xsession $HOME/.xsession
@@ -75,9 +84,4 @@ configure_dots() {
     echo "DONE!"
 }
 
-case "$1" in
-    --pre) pre_install ;;
-    --post) post_install ;;
-    --dots) configure_dots ;;
-esac
-
+configure_dots
